requirements/requirements.txt
coverage==4.2
Django==1.11.4
django-filter==1.0.4
djangorestframework==3.6.4
djangorestframework-serializer-extensions==0.5.2
dnspython==1.15.0
email-validator==1.0.2
factory-boy==2.9.2
Faker==0.8.1
hashids==1.2.0
idna==2.6
Markdown==2.6.9
mysqlclient==1.3.10
python-dateutil==2.6.1
pytz==2017.2
six==1.10.0
UkPostcodeParser==1.1.1
